Issue #237 Move Typeahead component in -extensions because it is not part of Twitter Bootstrap anymore
