--- bootstrap-core/src/main/java/de/agilecoders/wicket/core/markup/html/bootstrap/form/typeahead/IDataSource.java
@@ -1,19 +1,0 @@
-package de.agilecoders.wicket.core.markup.html.bootstrap.form.typeahead;
-
-import org.apache.wicket.util.io.IClusterable;
-
-import java.util.List;
-
-/**
- * Simple data source for e.g. {@link de.agilecoders.wicket.core.markup.html.bootstrap.form.typeahead.js.Typeahead}
- *
- * @author miha
- */
-public interface IDataSource<T> extends IClusterable {
-
-    /**
-     * @return a list of data of type T
-     */
-    List<T> load();
-
-}
