+++ bootstrap-samples/src/main/java/de/agilecoders/wicket/samples/pages/Javascript.java
@@ -12,10 +12,9 @@ import de.agilecoders.wicket.core.markup.html.bootstrap.carousel.CarouselImage;
 import de.agilecoders.wicket.core.markup.html.bootstrap.carousel.ICarouselImage;
 import de.agilecoders.wicket.core.markup.html.bootstrap.components.PopoverBehavior;
 import de.agilecoders.wicket.core.markup.html.bootstrap.components.TooltipBehavior;
-import de.agilecoders.wicket.core.markup.html.bootstrap.form.typeahead.Dataset;
-import de.agilecoders.wicket.core.markup.html.bootstrap.form.typeahead.Typeahead;
+import de.agilecoders.wicket.extensions.markup.html.bootstrap.form.typeahead.Dataset;
+import de.agilecoders.wicket.extensions.markup.html.bootstrap.form.typeahead.Typeahead;
 import de.agilecoders.wicket.core.markup.html.bootstrap.image.IconType;
-import de.agilecoders.wicket.core.markup.html.bootstrap.layout.col.LargeScreenSpanType;
 import de.agilecoders.wicket.core.markup.html.bootstrap.navbar.NavbarAjaxLink;
 import de.agilecoders.wicket.core.markup.html.bootstrap.tabs.AjaxLazyLoadTextContentTab;
 import de.agilecoders.wicket.core.markup.html.bootstrap.tabs.Collapsible;
